admin view dan update

+++ routes/web.php
@@ -40,8 +40,9 @@ Route::group(['middleware'=>'admin','auth'] ,function(){
 Route::get('/admin/participant', 'AdminController@home')->name('admin');
 Route::get('/admin/payment', 'AdminController@payment')->name('admin-payment');
 Route::get('/admin/view/{id}', 'AdminController@view')->name('admin-view');
-Route::delete('/admin/delete/{id}', 'AdminController@destroy')->name('delete');
+Route::delete('/admin/delete/{id}', 'AdminController@destroy')->name('admin-delete');
 Route::get('/admin/edit/{id}', 'AdminController@edit')->name('admin-edit');
 Route::patch('/admin/update/{id}', 'AdminController@store')->name('admin-update');
+Route::patch('/admin/updatemember/{id}', 'AdminController@storemember')->name('admin-updatemember');
 });
 Route::post('/sendemail/send', 'SendEmailController@send')->name('send');
